Revert "add new update Ultimate-HD-Skin for openATV 6.5 Python3"
This reverts commit 1c126faf93adac1cf853ee2a1c978f23b52a6558.

+++ Ultimate-HD-Skin-4ATV/usr/lib/enigma2/python/Components/Renderer/UltimateWatches.py
@@ -1,5 +1,5 @@
 import math
-from Components.Renderer.Renderer import Renderer
+from Renderer import Renderer
 from skin import parseColor
 from enigma import eCanvas, eSize, gRGB, eRect
 from Components.VariableText import VariableText
@@ -38,7 +38,7 @@ class UltimateWatches(Renderer):
     def hand(self):
         width = self.instance.size().width()
         height = self.instance.size().height()
-        r = int(min(width, height) / 2)
+        r = min(width, height) / 2
         endX, endY = self.calculate(self.numval, r, r)
         self.draw_line(r, r, endX, endY)
